Added harvester controller
